+++ src/integer_mod_q/mat_zq/arithmetic/add.rs
@@ -10,11 +10,14 @@
 
 use super::super::MatZq;
 use crate::error::MathError;
+use crate::integer::MatZ;
 use crate::macros::arithmetics::{
     arithmetic_trait_borrowed_to_owned, arithmetic_trait_mixed_borrowed_owned,
+    arithmetic_trait_reverse,
 };
 use crate::traits::MatrixDimensions;
-use flint_sys::fmpz_mod_mat::fmpz_mod_mat_add;
+use flint_sys::fmpz_mat::fmpz_mat_add;
+use flint_sys::fmpz_mod_mat::{_fmpz_mod_mat_reduce, fmpz_mod_mat_add};
 use std::ops::Add;
 
 impl Add for &MatZq {
@@ -48,6 +51,61 @@ impl Add for &MatZq {
         self.add_safe(other).unwrap()
     }
 }
+
+impl Add<&MatZq> for &MatZ {
+    type Output = MatZq;
+
+    /// Implements the [`Add`] trait for a [`MatZ`] and a [`MatZq`] matrix.
+    /// [`Add`] is implemented for any combination of [`MatZ`] and [`MatZq`] and vice versa.
+    ///
+    /// Parameters:
+    /// - `other`: specifies the value to add to `self`
+    ///
+    /// Returns the sum of both numbers as a [`MatZq`].
+    ///
+    /// # Examples
+    /// ```
+    /// use qfall_math::{integer::MatZ, integer_mod_q::MatZq};
+    /// use std::str::FromStr;
+    ///
+    /// let a = MatZ::from_str("[[1, 2, 3],[3, 4, 5]]").unwrap();
+    /// let b = MatZq::from_str("[[1, 9, 3],[1, 0, 5]] mod 7").unwrap();
+    ///
+    /// let c = &a + &b;
+    /// let d = a.clone() + b.clone();
+    /// let e = &b + &a;
+    /// let f = b + a;
+    /// ```
+    ///
+    /// # Panics ...
+    /// - if the dimensions of both matrices mismatch.
+    fn add(self, other: &MatZq) -> Self::Output {
+        if self.get_num_rows() != other.get_num_rows()
+            || self.get_num_columns() != other.get_num_columns()
+        {
+            panic!(
+                "Tried to add a '{}x{}' matrix and a '{}x{}' matrix.",
+                self.get_num_rows(),
+                self.get_num_columns(),
+                other.get_num_rows(),
+                other.get_num_columns()
+            );
+        }
+
+        let mut out = MatZq::new(self.get_num_rows(), self.get_num_columns(), other.get_mod());
+        unsafe {
+            fmpz_mat_add(&mut out.matrix.mat[0], &self.matrix, &other.matrix.mat[0]);
+            _fmpz_mod_mat_reduce(&mut out.matrix);
+        }
+        out
+    }
+}
+
+arithmetic_trait_borrowed_to_owned!(Add, add, MatZ, MatZq, MatZq);
+arithmetic_trait_mixed_borrowed_owned!(Add, add, MatZ, MatZq, MatZq);
+arithmetic_trait_reverse!(Add, add, MatZq, MatZ, MatZq);
+arithmetic_trait_borrowed_to_owned!(Add, add, MatZq, MatZ, MatZq);
+arithmetic_trait_mixed_borrowed_owned!(Add, add, MatZq, MatZ, MatZq);
 
 impl MatZq {
     /// Implements addition for two [`MatZq`] matrices.
@@ -197,5 +255,88 @@ mod test_add {
         assert!(a.add_safe(&b).is_err());
         assert!(c.add_safe(&b).is_err());
         assert!(a.add_safe(&d).is_err());
+    }
+}
+
+#[cfg(test)]
+mod test_add_matz {
+    use super::MatZq;
+    use crate::integer::MatZ;
+    use std::str::FromStr;
+
+    /// Ensures that addition between [`MatZ`] and [`MatZq`] works properly incl. reduction mod q.
+    #[test]
+    fn small_numbers() {
+        let a = MatZ::from_str("[[1, 2],[3, 4]]").unwrap();
+        let b = MatZq::from_str("[[5, 6],[9, 10]] mod 11").unwrap();
+        let cmp = MatZq::from_str("[[6, 8],[1, 3]] mod 11").unwrap();
+
+        let res_0 = &a + &b;
+        let res_1 = &b + &a;
+
+        assert_eq!(res_0, res_1);
+        assert_eq!(cmp, res_0);
+    }
+
+    /// Testing addition for large numbers
+    #[test]
+    fn large_numbers() {
+        let a: MatZ =
+            MatZ::from_str(&format!("[[1, 2, {}],[3, -4, {}]]", i64::MIN, i128::MAX,)).unwrap();
+        let b: MatZq = MatZq::from_str(&format!(
+            "[[1, 2, {}],[3, 9, {}]] mod {}",
+            i64::MIN + 1,
+            i128::MAX,
+            u64::MAX - 58
+        ))
+        .unwrap();
+
+        let c = a + &b;
+
+        assert_eq!(
+            c,
+            MatZq::from_str(&format!(
+                "[[2, 4, -{}],[6, 5, {}]] mod {}",
+                u64::MAX,
+                u128::MAX - 1,
+                u64::MAX - 58
+            ))
+            .unwrap()
+        );
+    }
+
+    /// Ensures that addition between [`MatZ`] and [`MatZq`] is available for any combination.
+    #[test]
+    fn available() {
+        let a = MatZ::new(2, 2);
+        let b = MatZq::new(2, 2, 7);
+
+        let _ = &b + &a;
+        let _ = &b + a.clone();
+        let _ = b.clone() + &a;
+        let _ = b.clone() + a.clone();
+        let _ = &a + &b;
+        let _ = &a + b.clone();
+        let _ = a.clone() + &b;
+    }
+
+    /// Ensures that mismatching rows results in a panic.
+    #[test]
+    #[should_panic]
+    fn mismatching_rows() {
+        let a = MatZ::new(3, 2);
+        let b = MatZq::new(2, 2, 7);
+
+        let _ = &b + &a;
+    }
+
+    /// Ensures that mismatching columns results in a panic.
+    #[test]
+    #[should_panic]
+    fn mismatching_column() {
+        let a = MatZ::new(2, 3);
+        let b = MatZq::new(2, 2, 7);
+
+        let _ = &b + &a;
     }
 }
